Changes to Dashboard files and Login-Register JS files
