--- ConsultApp/static/ConsultApp/dashboard.css
@@ -1,177 +1,0 @@
-body {
-  margin: 0;
-  font-family: Arial, sans-serif;
-  background: linear-gradient(120deg, #001f54, #e6f0ff); 
-  color: #333;
-}
-
-.container {
-  display: flex;
-  min-height: 100vh;
-}
-
-/* Sidebar */
-.sidebar {
-  width: 240px;
-  background: #ffffff;
-  color: #333;
-  min-height: 100vh;
-  padding: 25px 15px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-radius: 0 20px 20px 0; 
-  box-shadow: 4px 0 15px rgba(0,0,0,0.1); 
-  background: linear-gradient(180deg, #ffffff, #f0f4ff); 
-  position: relative;
-}
-
-
-
-
-.sidebar .logo {
-  text-align: center;
-  margin-bottom: 40px;
-}
-
-.sidebar .logo img {
-  width: 180px; 
-  border-radius: 8px;
-}
-
-.menu {
-  list-style: none;
-  padding: 0;
-  width: 100%;
-}
-
-.menu li {
-  padding: 12px 15px;
-  margin: 8px 0;
-  border-radius: 8px;
-  cursor: pointer;
-  font-weight: bold;
-  text-align: left;
-  transition: all 0.3s;
-  border: 1px solid transparent;
-}
-
-.menu li:hover {
-  background: #dce4ff;
-  border-color: #5d3a99;
-  color: #5d3a99;
-  box-shadow: 0 2px 6px rgba(0,0,0,0.1);
-}
-
-/* Main Content */
-.main-content {
-  flex: 1;
-  padding: 30px 50px;
-}
-
-/* Header */
-.header {
-  background: #f3f5ff;
-  padding: 25px 20px;
-  border-radius: 12px;
-  margin-bottom: 30px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.05);
-}
-
-.header h1 {
-  margin: 0;
-  color: #4b2f80;
-}
-
-.header p {
-  margin: 5px 0 0;
-  color: #555;
-}
-
-/* Stats */
-.stats {
-  display: flex;
-  gap: 20px;
-  margin-bottom: 30px;
-}
-
-.card {
-  flex: 1;
-  background: #ffffff;
-  padding: 20px;
-  border-radius: 12px;
-  text-align: center;
-  box-shadow: 0px 4px 12px rgba(0,0,0,0.08);
-  transition: transform 0.2s;
-}
-
-.card:hover {
-  transform: translateY(-4px);
-}
-
-.card h2 {
-  margin: 0;
-  font-size: 28px;
-}
-
-.card.purple { border-left: 6px solid #7e5ac3; }
-.card.blue { border-left: 6px solid #5a9bd5; }
-.card.green { border-left: 6px solid #38a638; }
-
-/* Tasks & Forms */
-.tasks h2 {
-  margin-bottom: 15px;
-  color: #4b2f80;
-}
-
-.task-card {
-  background: #ffffff;
-  padding: 20px;
-  margin-bottom: 20px;
-  border-radius: 12px;
-  box-shadow: 0px 4px 12px rgba(0,0,0,0.05);
-}
-
-/* Buttons & Inputs */
-button {
-  padding: 10px 18px;
-  background: #5D3A99;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  font-weight: bold;
-  transition: all 0.3s;
-}
-
-button:hover {
-  background: #7e5ac3;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.2);
-}
-
-input {
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  margin: 8px 0;
-  width: 100%;
-  transition: border-color 0.3s;
-}
-
-input:focus {
-  border-color: #5D3A99;
-  outline: none;
-}
-
-.verified-badge {
-  display: inline-block;
-  background: linear-gradient(135deg, #5D3A99, #7e5ac3);
-  color: white;
-  font-size: 12px;
-  font-weight: bold;
-  padding: 4px 10px;
-  border-radius: 6px;
-  margin-left: 10px;
-  vertical-align: middle;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.15);
-}
