--- ConsultApp/templates/ConsultApp/consultant-dashboard.html
@@ -1,85 +1,0 @@
-{% load static %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Consultant Dashboard</title>
-  <link rel="stylesheet" href="{% static 'ConsultApp/dashboard.css'%}">
-</head>
-<body>
-  <div class="container">
-    <aside class="sidebar">
-      <div class="logo">
-        <img src="{% static 'ConsultApp/Logo-Title.png'%}" alt="ResearchMate Logo">
-      </div>
-      <ul class="menu">
-        <li>🏠 Home</li>
-        <li>📘 Profile</li>
-        <li>⚙️ Settings</li>
-        <li>📅 Bookings</li>
-        <li>🚪 Logout</li>
-      </ul>
-    </aside>
-
-    <main class="main-content">
-      <header class="header">
-        <h1>
-          Welcome, <span id="consultantName">Consultant</span>
-          <span id="verifiedBadge" class="verified-badge" style="display:none;">✔ Verified</span>
-        </h1>
-        <p>Manage your profile and guide students in their research</p>
-      </header>
-
-      <section class="stats">
-        <div class="card purple">
-          <h2>3</h2>
-          <p>Active Students</p>
-        </div>
-        <div class="card blue">
-          <h2>5</h2>
-          <p>Consultations</p>
-        </div>
-        <div class="card green">
-          <h2>12</h2>
-          <p>Completed Projects</p>
-        </div>
-      </section>
-
-      <section class="tasks">
-        <h2>Consultant Actions</h2>
-        <button id="verifyBtn" class="primary-btn">Verify Yourself as Consultant</button>
-      </section>
-    </main>
-  </div>
-
-  <script>
-    const verifiedBadge = document.getElementById("verifiedBadge");
-    const verifyBtn = document.getElementById("verifyBtn");
-    const consultantName = document.getElementById("consultantName");
-
-    // Initially everything is unverified
-    let isVerified = false;
-    let name = "Consultant";
-
-    // Check if user is returning from verification page
-    const urlParams = new URLSearchParams(window.location.search);
-    if (urlParams.get("verified") === "true") {
-      isVerified = true;
-      name = urlParams.get("name") || "Consultant";
-    }
-
-    // Update dashboard display based on verification status
-    if (isVerified) {
-      consultantName.textContent = name;
-      verifiedBadge.style.display = "inline-block";
-      verifyBtn.style.display = "none"; // Hide the verification button
-    }
-
-    // Redirect to verification page on button click
-    verifyBtn.addEventListener("click", () => {
-      window.location.href = "{% url 'consultant_verification' %}";
-    });
-  </script>
-</body>
-</html>
